+++ compare-frontend/src/app/modules/wizard/converters/eqts/3_exposure.ts
@@ -1,0 +1,33 @@
+import { ScenarioName, RiesgosScenarioState, RiesgosProductResolved } from "src/app/state/state";
+import { Converter } from "../../converter.service";
+import { WizardComposite } from "../../wizard.service";
+
+
+export class Exposure implements Converter {
+    applies(scenario: ScenarioName, step: string): boolean {
+        return step === "Exposure";
+    }
+
+    getInfo(state: RiesgosScenarioState, data: RiesgosProductResolved[]): WizardComposite {
+        const step = state.steps.find(s => s.step.id === "Exposure")!;
+        const inputProd = step.step.inputs[0];
+        const currentValue = state.products.find(p => p.id === inputProd.id)?.value;
+
+        return {
+            hasFocus: false,
+            inputs: [{
+                formtype: "string-select",
+                label: "exposure model",
+                productId: inputProd.id,
+                options: Object.fromEntries(inputProd.options!.map(v => [v, v])),
+                currentValue: currentValue
+            }],
+            step: step,
+            // legend: {
+            //     component: fdsafs
+            //     args: {}
+            // }
+        }
+    }
+
+}